Refactor test-setup to use factory approach rather than classes
